08. TiendaStark2023/TiendaStark/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cx-Oracle==8.3.0
Django==4.2
djangorestframework==3.14.0
idna==3.4
marshmallow==3.17.0
packaging==23.1
Pillow==9.5.0
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
transbank-sdk==4.0.0
tzdata==2023.3
urllib3==1.26.15
